\name{perDiag_C}
\alias{EM}
\title{perDiag}
\usage{
perDiag(data)

}

\arguments{
  \item{data}{A dataset}

}
\description{
perDiag takes you into a menu-input, user-programme interaction. The objective is to provide easy access to various statistical tools for investigation on data diagnostics with linear regression.
}
\examples{
perDiag(mtcars)
}
